requirements.txt
selenium==3.141.0
beautifulsoup4==4.10.0
pymongo==4.0.1
python-dotenv==0.19.0
